Almsot ready for use

+++ backend-bugfinder/src/main/bbf/com/stepanov/bbf/executor/compilers/JVMCompiler.kt
@@ -53,8 +53,8 @@ class JVMCompiler(private val arguments: String = "") : CommonCompiler() {
         val threadPool = Executors.newCachedThreadPool()
         val trashDir = "tmp/trash/"
         //Clean dir
-        FileUtils.cleanDirectory(File(trashDir))
-
+        if (File(trashDir).exists())
+            FileUtils.cleanDirectory(File(trashDir))
         val args =
                 if (arguments.isEmpty())
                     "$pathToFile -d $trashDir".split(" ")
